+++ cds-feature-attachments/src/test/java/com/sap/cds/feature/attachments/service/AttachmentsServiceImplTest.java
@@ -77,7 +77,7 @@ class AttachmentsServiceImplTest {
 	}
 
 	@ParameterizedTest
-	@ValueSource(booleans = {true, false})
+	@ValueSource(booleans = { true, false })
 	@NullSource
 	void createAttachmentInsertsData(Boolean isExternalCreated) {
 		var contextReference = new AtomicReference<AttachmentCreateEventContext>();
@@ -93,7 +93,8 @@ class AttachmentsServiceImplTest {
 		serviceSpi.on(AttachmentService.EVENT_CREATE_ATTACHMENT, "", handler);
 		var stream = mock(InputStream.class);
 		Map<String, Object> ids = Map.of("ID1", "value1", "id2", "Value2");
-		var input = new CreateAttachmentInput(ids, mock(CdsEntity.class), "fileName", "mimeType", stream);
+		var input = new CreateAttachmentInput(ids, mock(CdsEntity.class), "fileName", "mimeType", stream, Map.of(),
+				null);
 
 		var result = cut.createAttachment(input);
 
@@ -116,7 +117,8 @@ class AttachmentsServiceImplTest {
 		}).when(handler).process(any());
 		serviceSpi.on(AttachmentService.EVENT_CREATE_ATTACHMENT, "", handler);
 		Map<String, Object> ids = Map.of("ID1", "value1", "id2", "Value2");
-		var input = new CreateAttachmentInput(ids, mock(CdsEntity.class), "fileName", "mimeType", mock(InputStream.class));
+		var input = new CreateAttachmentInput(ids, mock(CdsEntity.class), "fileName", "mimeType",
+				mock(InputStream.class), Map.of(), null);
 
 		var result = cut.createAttachment(input);